* Introduce skeletal ProgramFrontend * Make ImagePreprocessingParameters Frontend-agnostic * Split compileSpecificationInternal into versions that do and don't support updatable models. * Correct initialization and sharing of program-level scope * Add IRProgram::GetFunction() * Add Frontend method to tell tell if the underlying NN spec version is v1 or v2 * Add skeletal IR-based Espresso layer generator * Add temporary NN spec v2-only path to run IR-based Espresso layer translator * IRScope throws on failed search * Add IRValue::CopyTo() * Add GetNumElements, As and TryAs to IRValueType * First cut at implementing InnerProduct * Implement reading weights from file * Stub implementations of GetOutputNames and CollectImagePreprocessingParameters in ProgramFrontend * enum for nonlinearity types and string -> enum converter * Implement enough of InnerProduct and Activation to build a simple network * Unit test to exercise Program --> Espresso --> execution path * Port very basic layer and output shape validation * Resolve some easy TODO items * Make testCompileAndLoadProgram with validation code * Add radar for ProgramFrontend::IsNdRequired * Fix merge issue

+++ mlmodel/src/ILIL/Program/ProgramIRValueType.cpp
@@ -16,7 +16,7 @@ static std::unique_ptr<IRDimension> ParseDimension(const V5::Dimension& specDim)
 {
     switch (specDim.dimension_case()) {
         case V5::Dimension::kSize:
-            return std::make_unique<IRConstantDimension>(specDim.size());
+            return std::make_unique<IRConstantDimension>(static_cast<uint64_t>(specDim.size()));
         case V5::Dimension::kSymbol:
             return std::make_unique<IRSymbolicDimension>(specDim.symbol());
         case V5::Dimension::DIMENSION_NOT_SET:
